+++ CHANGELOG.markdown
@@ -1,3 +1,9 @@
+0.9.0
+-----
+* Add `generalBracket` to the `MonadMask` typeclass, allowing more
+  valid instances
+* Add `MonadMask` instances for `ExceptT` and `ErrorT`
+
 0.8.3
 -----
 * `MonadCatch` and `MonadMask` instances for `Either SomeException`